Stripe patches
- Add handling for Stripe API Key handling @lcampanis
